Move get_siret_token
